dermacare/backend/requirements.txt
Flask
flask-cors
flask-pymongo
PyJWT
Werkzeug
Pillow
numpy
